+++ contrib/git-changes-action/detector/module/suite_test.go
@@ -1,4 +1,4 @@
-package detector_test
+package moduledetector_test
 
 import (
 	"github.com/Flaque/filet"