+++ packages/contracts-communication/test/legacy/MessageBus.Management.t.sol
@@ -1,6 +1,8 @@
 // SPDX-License-Identifier: MIT
 pragma solidity 0.8.20;
 
+import {LegacyMessageLib} from "../../contracts/legacy/libs/LegacyMessage.sol";
+
 import {MessageBusBaseTest, InterchainClientV1Mock} from "./MessageBus.Base.t.sol";
 
 // solhint-disable func-name-mixedcase
@@ -9,8 +11,10 @@ contract MessageBusManagementTest is MessageBusBaseTest {
     uint256 public constant LENGTH_ESTIMATE = 100;
 
     function mockInterchainFees(uint256 length) internal {
+        uint256 legacyMessageLen = LegacyMessageLib.payloadSize(length);
         bytes memory expectedCalldata = abi.encodeCall(
-            InterchainClientV1Mock.getInterchainFee, (REMOTE_CHAIN_ID, execService, icModules, icOptions, length)
+            InterchainClientV1Mock.getInterchainFee,
+            (REMOTE_CHAIN_ID, execService, icModules, icOptions, legacyMessageLen)
         );
         vm.mockCall(icClient, expectedCalldata, abi.encode(MOCK_FEE));
     }